Moved lifters to their own file <lifters.h>. This avoids the circular include issues in a much more elegant way than reshuffling <parcel.h> the way I did earlier
