+++ examples/C++/main.cpp
@@ -1,6 +1,7 @@
 
 #include <SHARPlib/constants.h>
 #include <SHARPlib/layer.h>
+#include <SHARPlib/lifters.h>
 #include <SHARPlib/parcel.h>
 #include <SHARPlib/profile.h>
 #include <SHARPlib/thermo.h>
