+++ src/SHARPlib/params/convective.cpp
@@ -14,6 +14,7 @@
 
 #include <SHARPlib/constants.h>
 #include <SHARPlib/layer.h>
+#include <SHARPlib/lifters.h>
 #include <SHARPlib/parcel.h>
 #include <SHARPlib/thermo.h>
 #include <SHARPlib/winds.h>
